move Param structs to main db module
